commentary/requirements.txt
pandas
jinja2
pydantic
matplotlib
numpy
python-dateutil
xlsxwriter